Convert to use touchbar-registry
