+++ keymaps/touchbar.json
@@ -1,5 +1,2 @@
 {
-  "atom-workspace": {
-    "ctrl-alt-o": "touchbar:toggle"
-  }
 }
