+++ lib/touchbar-edit-view.js
@@ -111,6 +111,7 @@ export default class TouchBarEditView {
 
   remove(i) {
     const oldConfig = JSON.parse(atom.config.get('touchbar.buttons'))
+    const id = oldConfig[i].id;
     oldConfig.splice(i, 1)
     atom.config.set('touchbar.buttons', JSON.stringify(oldConfig))
     etch.update(this)
